backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
sqlalchemy
pydantic
python-dotenv
httpx
